+++ superfreetts_addon/config_models.py
@@ -379,6 +379,7 @@ class TextProcessing:
         self.strip_brackets = constants.TEXT_PROCESSING_DEFAULT_STRIP_BRACKETS
         self.strip_cloze = constants.TEXT_PROCESSING_DEFAULT_STRIP_CLOZE
         self.ssml_convert_characters = constants.TEXT_PROCESSING_DEFAULT_SSML_CHARACTERS
+        self.sanitize_special_characters = constants.TEXT_PROCESSING_DEFAULT_SANITIZE_SPECIAL_CHARACTERS
         self.run_replace_rules_after = constants.TEXT_PROCESSING_DEFAULT_REPLACE_AFTER
         self.ignore_case = constants.TEXT_PROCESSING_DEFAULT_IGNORE_CASE
 
@@ -406,6 +407,7 @@ class TextProcessing:
             'strip_brackets': self.strip_brackets,
             'strip_cloze': self.strip_cloze,
             'ssml_convert_characters': self.ssml_convert_characters,
+            'sanitize_special_characters': self.sanitize_special_characters,
             'run_replace_rules_after': self.run_replace_rules_after,
             'ignore_case': self.ignore_case,
             'text_replacement_rules': [x.serialize() for x in self.text_replacement_rules]